+++ CHANGELOG.md
@@ -15,6 +15,10 @@ incremented for features.
 
 * ts: Allow preloading instructions for state rpc transactions ([cf9c84](https://github.com/project-serum/anchor/commit/cf9c847e4144989b5bc1936149d171e90204777b)).
 
+## Fixes
+
+* cli: Propagates mocha test exit status on error.
+
 ## [0.2.1] - 2021-02-11
 
 ### Features